+++ .makefiles/bump_blocklet_version.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 
 NEW_VERSION=$(cat version)
-abtnode blocklet:version $NEW_VERSION
+blocklet version $NEW_VERSION
 git add blocklet.yml